% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mars_id}
\alias{mars_id}
\title{Product IDs and site names of Mars Science Laboratory XRPD data}
\format{
A data frame
\describe{
\item{SITE_NAME}{The name of the site sampled on Mars}
\item{PRODUCT_ID}{The unique identifier for the processed XRPD data}
\item{SAMPLE_TYPE}{Either 'Scoop' or 'Drill`}
\item{SOL_START}{The sol (i.e. Martian day number) that data upload started}
\item{SOL_END}{The sol (i.e. Martian day number) that data upload ended}
}
}
\usage{
mars_id
}
\description{
Product IDs and site names of 31 XRPD patterns recorded on the Mars Science Laboratory
aboard NASA's Curiosity Rover. Data were obtained from the Geosciences Data Volume
Online (https://pds-geosciences.wustl.edu/msl/msl-m-chemin-4-rdr-v1/) on 14.07.21.
}
\keyword{datasets}
