% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mars_xrpd}
\alias{mars_xrpd}
\title{Martian XRPD data from NASA's Mars Science Laboratory}
\format{
A multiXY object
}
\usage{
mars_xrpd
}
\description{
A multiXY object of 31 XRPD patterns recorded on the Mars Science Laboratory
aboard NASA's Curiosity Rover. Data were obtained from the Geosciences Data Volume
Online (https://pds-geosciences.wustl.edu/msl/msl-m-chemin-4-rdr-v1/) on 14.07.21.
}
\keyword{datasets}
