% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_locations}
\alias{scotland_locations}
\title{Profile locations of the Scottish soil samples}
\format{
A data frame
}
\usage{
scotland_locations
}
\description{
Sample IDs, profile IDs, profile longitude and profile latitude of the
Scottish soil samples.
}
\keyword{datasets}
