% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_xrpd}
\alias{scotland_xrpd}
\title{Scottish XRPD data}
\format{
A data frame
}
\usage{
scotland_xrpd
}
\description{
A data frame of 775 columns containing XRPD data. First column is the
2theta axis, followed by 774 columns of count intensities, with sample IDs
as column names.
}
\keyword{datasets}
