% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mars_id}
\alias{mars_id}
\title{Product IDs and site names of Mars Science Laboratory XRPD data}
\format{
A data frame
}
\usage{
mars_id
}
\description{
Product IDs and site names of 31 XRPD patterns recorded on the Mars Science Laboratory
aboard NASA's Curiosity Rover. Data were obtained from the Geosciences Data Volume
Online (https://pds-geosciences.wustl.edu/msl/msl-m-chemin-4-rdr-v1/) on 14.07.21.
}
\keyword{datasets}
