% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_locations}
\alias{scotland_locations}
\title{Profile locations of the Scottish soil samples}
\format{
A data frame
\describe{
\item{SAMPLE_ID}{The unique identifier of the sampled soil horizon}
\item{PROFILE_ID}{The unique identifier of the soil profile}
\item{PROFILE_LONGITUDE}{Longitude of the soil profile}
\item{PROFILE_LATITUDE}{Latitude of the soil profile}
\item{DATASET}{Denotes the dataset that the sample is associated with}
}
}
\usage{
scotland_locations
}
\description{
Sample IDs, profile IDs, profile longitude and profile latitude of the
Scottish soil samples.
}
\keyword{datasets}
